+++ test/plugin-minimizer-option.test.js
@@ -454,12 +454,8 @@ describe("plugin minify option", () => {
   });
 
   it("should transform image to webp", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "filename-1");
     const stats = await runWebpack({
       entry: path.join(fixturesPath, "./empty-entry.js"),
-      output: {
-        path: outputDir,
-      },
       emitPlugin: true,
       emitPluginOptions: { fileNames: ["./nested/deep/plugin-test.png"] },
       imageminPluginOptions: {
@@ -478,7 +474,7 @@ describe("plugin minify option", () => {
     );
     const file = path.resolve(
       __dirname,
-      outputDir,
+      compilation.options.output.path,
       "./nested/deep/plugin-test.webp"
     );
     const ext = await fileType.fromFile(file);
@@ -490,12 +486,8 @@ describe("plugin minify option", () => {
   });
 
   it("should transform image to webp with flat filename", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "filename-2");
     const stats = await runWebpack({
       entry: path.join(fixturesPath, "./empty-entry.js"),
-      output: {
-        path: outputDir,
-      },
       emitPlugin: true,
       emitPluginOptions: { fileNames: ["./nested/deep/plugin-test.png"] },
       imageminPluginOptions: {
@@ -512,7 +504,11 @@ describe("plugin minify option", () => {
     const transformedAssets = Object.keys(assets).filter((asset) =>
       asset.includes("plugin-test.webp")
     );
-    const file = path.resolve(__dirname, outputDir, "plugin-test.webp");
+    const file = path.resolve(
+      __dirname,
+      compilation.options.output.path,
+      "plugin-test.webp"
+    );
     const ext = await fileType.fromFile(file);
 
     expect(/image\/webp/i.test(ext.mime)).toBe(true);
@@ -522,12 +518,8 @@ describe("plugin minify option", () => {
   });
 
   it("should transform image to webp with nested filename", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "filename-3");
     const stats = await runWebpack({
       entry: path.join(fixturesPath, "./empty-entry.js"),
-      output: {
-        path: outputDir,
-      },
       emitPlugin: true,
       emitPluginOptions: { fileNames: ["plugin-test.png"] },
       imageminPluginOptions: {
@@ -546,7 +538,7 @@ describe("plugin minify option", () => {
     );
     const file = path.resolve(
       __dirname,
-      outputDir,
+      compilation.options.output.path,
       "./nested/deep/plugin-test.webp"
     );
     const ext = await fileType.fromFile(file);
@@ -558,12 +550,8 @@ describe("plugin minify option", () => {
   });
 
   it("should transform image to webp with filename pointing to other directory", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "filename-4");
     const stats = await runWebpack({
       entry: path.join(fixturesPath, "./empty-entry.js"),
-      output: {
-        path: outputDir,
-      },
       emitPlugin: true,
       emitPluginOptions: { fileNames: ["./nested/deep/plugin-test.png"] },
       imageminPluginOptions: {
@@ -580,7 +568,11 @@ describe("plugin minify option", () => {
     const transformedAssets = Object.keys(assets).filter((asset) =>
       asset.includes("./other/plugin-test.webp")
     );
-    const file = path.resolve(__dirname, outputDir, "./other/plugin-test.webp");
+    const file = path.resolve(
+      __dirname,
+      compilation.options.output.path,
+      "./other/plugin-test.webp"
+    );
     const ext = await fileType.fromFile(file);
 
     expect(/image\/webp/i.test(ext.mime)).toBe(true);
@@ -590,12 +582,8 @@ describe("plugin minify option", () => {
   });
 
   it("should transform image to webp with filename when filename is function", async () => {
-    const outputDir = path.resolve(__dirname, "outputs", "filename-4");
     const stats = await runWebpack({
       entry: path.join(fixturesPath, "./empty-entry.js"),
-      output: {
-        path: outputDir,
-      },
       emitPlugin: true,
       emitPluginOptions: { fileNames: ["./nested/deep/plugin-test.png"] },
       imageminPluginOptions: {
@@ -612,7 +600,11 @@ describe("plugin minify option", () => {
     const transformedAssets = Object.keys(assets).filter((asset) =>
       asset.includes("./other/plugin-test.webp")
     );
-    const file = path.resolve(__dirname, outputDir, "./other/plugin-test.webp");
+    const file = path.resolve(
+      __dirname,
+      compilation.options.output.path,
+      "./other/plugin-test.webp"
+    );
     const ext = await fileType.fromFile(file);
 
     expect(/image\/webp/i.test(ext.mime)).toBe(true);
